+++ src/lib/platform.cpp
@@ -6,14 +6,14 @@ bool FileExists(const char *filePath) {
 	return (dwAttrib != INVALID_FILE_ATTRIBUTES && !(dwAttrib & FILE_ATTRIBUTE_DIRECTORY));
 }
 
-void GetScreenResolution(size_t& width, size_t& height) {
+void GetScreenResolution(int& width, int& height) {
 	RECT desktop;
 	const HWND hDesktop = GetDesktopWindow();
 	GetWindowRect(hDesktop, &desktop);
 	width = desktop.right;
 	height = desktop.bottom;
 }
-#elif defined(linux) || defined(__linux__) || defined(__unix__)
+#elif defined(__APPLE__) || defined(linux) || defined(__linux__) || defined(__unix__)
 #include <fcntl.h>
 #include <unistd.h>
 #include <X11/Xlib.h>
@@ -24,7 +24,7 @@ bool FileExists(const char *filePath) {
 	return (access(filePath, F_OK ) != -1);
 }
 
-void GetScreenResolution(size_t& width, size_t& height) {
+void GetScreenResolution(int& width, int& height) {
 	Display* d = XOpenDisplay(NULL);
 	Screen* s = DefaultScreenOfDisplay(d);
 	width = s->width;
